Fix the Great Branch Deletion Disaster
I messed up big and deleted a branch locally on accident, and apparently
none of those commits (6 months worth or so) never actually made it to
Github's servers somehow?

Luckily my laptop had the latest changes on it, so I was able to quickly
turn on and offline the laptop, and then move the files back to my PC to
commit here.

0/10 not fun to recover deleted projects would not recommend.