Set up the virtual audio cable installation
Put in place some code to install the VAC from the exe. Seems to work
fine if the user cancels the installation and if it isn't installed,
need to test what happens when you go through with the installation.

Set up some logic to initialize and reset the audio devices, but that is
untested/not fully implemented until the VAC installation is done