VAC installs and is detected on launch.
Added logging for this. I've been tinkering with the VAC outside of this application and can't seem to get it to work correctly. I'll try again to get sound to output to both headphones/speakers and to the audio out/microphone but i may switch to trying the NAudio API. I added the latest version to the project.